# api/public.openapi.yaml
openapi: 3.0.3
info:
  title: Pratu Public API
  version: 0.2.0
  description: |
    The tenant-facing API. Every request is addressed to a tenant hostname
    (`{slug}.{base_domain}` or a custom domain): the Host header selects the
    tenant, and each tenant is its own OIDC issuer.

    **Flows** come in two variants. *API flows* (`/api` creation endpoints)
    are for native apps: they return `session_token` values and need no
    CSRF. *Browser flows* (`/browser` creation endpoints) set an HTTP-only
    CSRF cookie, return a `csrf_token` bound to the flow, require it on
    every submission, and set the `pratu_session` cookie instead of
    returning a session token.

    **Browser flows are content-negotiated.** A client that prefers
    `text/html` — or posts `application/x-www-form-urlencoded` — is driven
    by 303 redirects to the tenant's configured screens: flow creation
    lands on the screen with `?flow=`, a failed submission returns there
    with its messages persisted on the flow, a failure with no flow left
    to return to lands on the error screen with `?code=`, and a completed
    flow lands on `return_to` (validated against the tenant's allow-list)
    or the tenant's default return URL. A client that asks for
    `application/json` gets the JSON responses documented below,
    unchanged. Screens re-read the flow with
    `GET /self-service/flows/{id}`.

    Cookie-authenticated requests to state-changing session endpoints
...

...
  license:
...
servers:
...
paths:
...
  /self-service/registration/browser:
    get:
      tags: [registration]
      summary: Create a registration browser flow (sets the CSRF cookie)
      parameters:
        - { $ref: "#/components/parameters/schemaName" }
        - { $ref: "#/components/parameters/returnTo" }
      responses:
        "200": { $ref: "#/components/responses/Flow" }
        "303": { $ref: "#/components/responses/ScreenRedirect" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /self-service/login/browser:
    get:
      tags: [login]
      summary: Create a login browser flow (sets the CSRF cookie)
      parameters: [{ $ref: "#/components/parameters/returnTo" }]
      responses:
        "200": { $ref: "#/components/responses/Flow" }
        "303": { $ref: "#/components/responses/ScreenRedirect" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /self-service/flows/{id}:
    get:
      tags: [login]
      summary: Read a browser flow
      description: |
        What the tenant's screen renders after a redirect: the step the
        flow waits on (`state`), the fields to show, the second-factor
        methods available, and the messages from the last submission.
        Readable only by the browser that created the flow (its CSRF
        cookie must match); API flows are not exposed here.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { $ref: "#/components/responses/Flow" }
        "404": { $ref: "#/components/responses/Error" }
...
  /self-service/recovery/browser:
    get:
      tags: [recovery]
      summary: Create a recovery browser flow (sets the CSRF cookie)
      parameters: [{ $ref: "#/components/parameters/returnTo" }]
      responses:
        "200": { $ref: "#/components/responses/Flow" }
        "303": { $ref: "#/components/responses/ScreenRedirect" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /self-service/social/{provider}/browser:
    get:
      tags: [social]
      summary: Start a social sign-in round trip
      description: |
        302-redirects the browser to the provider with the flow ID as the
        OAuth2 state. On return, the callback logs in a known social
        account, links via provider-verified email matching a verified
        address, or registers a fresh passwordless identity — then either
        sets the session cookie and redirects to the tenant's
        `social_return_url`, or (with second factors enrolled) redirects
        with `mfa_flow`, `methods`, and `mfa_csrf` parameters for
        completion via the login TOTP/SMS endpoints.
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
      responses:
        "302": { description: Redirect to the provider }
        "404": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    sessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Opaque session token (also accepted as `Authorization Bearer`)
    sessionCookie:
      type: apiKey
      in: cookie
      name: pratu_session
      description: Browser-flow session cookie; state-changing calls also need `X-CSRF-Token`
  parameters:
...
    returnTo:
      name: return_to
      in: query
      required: false
      description: |
        Where to send the browser when the flow completes. Must be
        same-origin, share an origin with one of the tenant's configured
        screens, or match `ui.allowed_return_urls`; anything else is
        rejected with 400.
      schema: { type: string }
...
    schemaName:
      name: schema
      in: query
      required: false
      description: Identity Schema name (default "default")
      schema: { type: string }
...
  responses:
...
    Flow:
      description: A Self-Service Flow
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Flow" }
    ScreenRedirect:
      description: |
        HTML clients only: the browser is sent to the tenant's screen
        (`?flow=`), to the error screen (`?code=`), or to the flow's
        return URL.
      headers:
        Location:
          schema: { type: string, format: uri }
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    RateLimited:
      description: Rate limited; Retry-After indicates when the window rolls over
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
  schemas:
    APIError:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            details:
              type: array
              items: { type: string }
    Flow:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [registration, login, recovery, verification] }
        expires_at: { type: string, format: date-time }
        state:
          type: string
          enum: [choose_method, mfa_required, code_required, second_factor_required, password_required]
          description: The step the flow waits on
        messages:
          type: array
          description: What the screen should say, carried across a redirect
          items:
            type: object
            properties:
              type: { type: string, enum: [error, info, success] }
              text: { type: string }
              details:
                type: array
                items: { type: string }
        csrf_token: { type: string, description: Present on browser flows only }
        ui:
          type: object
          properties:
            fields:
              type: array
              items:
                type: object
                properties:
                  name: { type: string }
                  type: { type: string }
                  title: { type: string }
                  required: { type: boolean }
            methods:
              type: array
              description: Second factors this flow can be continued with
              items: { type: string, enum: [totp, sms] }
...